Restructuring the folder structure to align with the system flow.
