+++ src/ZooManagementSystem/DomainModels/Colour.java
@@ -1,4 +1,4 @@
-package ZooManagementSystem.Enums;
+package ZooManagementSystem.DomainModels;
 
 public enum Colour {
 	BLACK, GOLD, ORANGE, YELLOW, WHITE, GREEN, BLUE, BROWN, RED, CYAN
