+++ src/ZooManagementSystem/DomainModels/DogType.java
@@ -1,4 +1,4 @@
-package ZooManagementSystem.Enums;
+package ZooManagementSystem.DomainModels;
 
 public enum DogType {
     Akita,Bulldog,Poodle,Terriers
