+++ src/ZooManagementSystem/DomainModels/Gender.java
@@ -1,4 +1,4 @@
-package ZooManagementSystem.Enums;
+package ZooManagementSystem.DomainModels;
 
 public enum Gender {
     Male,Female;
