+++ src/ZooManagementSystem/DomainModels/Pattern.java
@@ -1,4 +1,4 @@
-package ZooManagementSystem.Enums;
+package ZooManagementSystem.DomainModels;
 
 public enum Pattern {
     DOTS, STRIPES, SPOTS, CLEAR
