+++ src/ZooManagementSystem/DomainModels/Tiger.java
@@ -1,6 +1,4 @@
-package ZooManagementSystem.Animals;
-
-import ZooManagementSystem.Enums.Gender;
+package ZooManagementSystem.DomainModels;
 
 public class Tiger extends CarnivorousAnimal{
         public Tiger(String name,int age, double weight,Gender gender){
