+++ src/ZooManagementSystem/Views/Main.java
@@ -1,7 +1,7 @@
 package ZooManagementSystem.Views;
 
-import ZooManagementSystem.Exceptions.AgeException;
-import ZooManagementSystem.Exceptions.HeightException;
+import ZooManagementSystem.Utils.AgeException;
+import ZooManagementSystem.Utils.HeightException;
 
 public class Main {
     public static void main(String[] args) throws HeightException, AgeException {
